slack_webhook_handler/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-secret-manager
google-cloud-tasks
google-cloud-firestore